+++ theme-generator/themes/sakura-day/variables.ts
@@ -1364,6 +1364,16 @@ export default {
       },
     },
   },
+  validationTooltip: {
+    error: {
+      backgroundColor: mixWithBackground(0.75, palette.red),
+      borderColor: mixWithBackground(0.5, palette.red),
+    },
+    warning: {
+      backgroundColor: mixWithBackground(0.75, palette.yellow),
+      borderColor: mixWithBackground(0.5, palette.yellow),
+    },
+  },
   vcs: {
     log: {
       commit: {