+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 
 ### Fixed
 - Fixed inactive text in text areas (e.g. GitHub copilot tool window).
+- Fixed validation tooltip colors.
 
 ## [0.11.0] - 2023-12-27
 ### Changed